remove diff from account info
