+++ lib/src/web/gigya_flutter_plugin_web.dart
@@ -18,7 +18,6 @@ import 'static_interop/parameters/registration.dart';
 import 'static_interop/response/registration_response.dart';
 import 'static_interop/response/response.dart';
 import 'static_interop/window.dart';
-import 'web_account_delegate.dart';
 import 'web_interruption_resolver.dart';
 
 /// An implementation of [GigyaFlutterPluginPlatform] that uses JavaScript static interop.
@@ -34,8 +33,6 @@ class GigyaFlutterPluginWeb extends GigyaFlutterPluginPlatform {
   InterruptionResolverFactory get interruptionResolverFactory {
     return const WebInterruptionResolverFactory();
   }
-
-  final WebAccountDelegate _accountDelegate = const WebAccountDelegate();
 
   @override
   Future<Map<String, dynamic>> addConnection(
@@ -151,15 +148,6 @@ class GigyaFlutterPluginWeb extends GigyaFlutterPluginPlatform {
     );
 
     return initRegistrationCompleter.future;
-  }
-
-  @override
-  Future<Map<String, dynamic>> getAccount({
-    bool invalidate = false,
-    Map<String, dynamic> parameters = const <String, dynamic>{},
-  }) async {
-    // Apparently, web does not support the invalidate argument.
-    return _accountDelegate.getAccount(parameters: parameters);
   }
 
   @override
@@ -447,10 +435,5 @@ class GigyaFlutterPluginWeb extends GigyaFlutterPluginPlatform {
     );
 
     return completer.future;
-  }
-
-  @override
-  Future<Map<String, dynamic>> setAccount(Map<String, dynamic> account) {
-    return _accountDelegate.setAccount(account);
   }
 }
