+++ lib/src/web/static_interop/parameters/conflicting_account.dart
@@ -1,0 +1,18 @@
+import 'dart:js_interop';
+
+import '../response/response.dart';
+
+/// This extension type represents the parameters for the `Accounts.getConflictingAccount` method.
+@JS()
+@anonymous
+@staticInterop
+extension type ConflictingAccountParameters._(JSObject _) implements JSObject {
+  /// Create a [ConflictingAccountParameters] instance using the given [callback] and [regToken].
+  /// 
+  /// The [callback] receives a [ConflictingAccountResponse] as argument,
+  /// and has [JSVoid] as return type.
+  external factory ConflictingAccountParameters({
+    JSFunction callback,
+    String regToken,
+  });
+}
