+++ lib/src/web/static_interop/response/response.dart
@@ -1,6 +1,5 @@
 import 'dart:js_interop';
 
-export 'account_response.dart';
 export 'conflicting_account_response.dart';
 export 'login_response.dart';
 export 'social_connection_response.dart';
